+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## 0.2.0 (?)
+ - [feature] travis / coveralls / pypi integration
+ - [feature] credential config files
+
 ## 0.1.0 (2014-09-09)
  - [feature] ConsumerKey lifecycle
  - [feature] OVH and RunAbove support